Acquire a VCL reference per named director
The API turned out a bit different in Varnish Cache, and for instance
references can only be acquired on the CLI thread. It prevents conflicts
with temperature changes that can only happen on the same thread. It
makes the handling simple and more balanced since the start function
acquires a reference that is released in the stop function.

I also managed to get named references in place, so instead of a simple
reference counter it's possible to let the user know what's holding the
cool-down.

Varnish 4.1.1 or later is needed for this change.