Avoid deadlocking when going cold.
When we receive a cold/discard vmod event, we do a pthread_join with
obj->mtx held.

This creates the possibility for a deadlock if, at the same time we're
about to pthread_join, another thread is in dynamic_update, which tries
to hold the same mutex:

vmod_event                         thread_A
obj->active = 0
lock(obj->mtx)
.
.                                  lock(obj->mtx)       -> deadlocks
.                                  .
pthread_join(thread_A)             .
.                                  .
.                                  unlock(obj->mtx)
.
unlock(obj->mtx)

The solution is after we get the DNS results, we do an early check if
the obj is still active. If it's not, just exit there and finish the
thread.