Remove spurious lock operations
The deadlock previously avoided with an early check may still exist.
Removing the offending lock operation should get rid of it for good
and even permit lookup threads to perform the last backend update, if
any, allowing a COLD director to resume with the latest available
lookup results if it goes WARM again.