Never allow a NULL reference in dom->current
If dom->current is the very last position on the queue and we're going
to delete it, VTAILQ_NEXT is going to be NULL. When this happens,
go back to the first element.

Also, when looking up a healthy backend, the dom->current can't be NULL
because we never allow next NULL, which would trigger an infinite
loop as well.

This fixes issue #7