Proper error handling in the constructor
Durations can't be negative, so it must be asserted. Illegal parameters
should fail vcl_init{}, but the VMOD still needs to allocate a proper
object. Otherwise it would be fed a NULL one in its destructor, that
currently expects a proper object, but more importantly method calls in
VCL would crash Varnish.