Add a sharing scope argument, per backend host_header default
Previously, backends were always shared at the director level. For
example, if d1.backend("foo") and d1.backend("bar") resolved to the
same address, only one backend was created for that address.

Using the share = HOST parameter, it is now possible to limit sharing
to backends of the same hostname (host argument to the backend()
method, which defaults to the Host: header).

Also, with share = HOST, the host_header parameter of backends is now
set to the host argument of the backend() method, which, in turn, gets
sent with probe requests unless overridden with a request attribute of
the probe or using the host_header argument during initialization of a
dynamic director.