add per .backend() port argument
With this change, it now becomes possible to specify the port to connect
to with each .backend() call. If given, the argument overrides the port
passed to the object constructor.

This also changes all logging output to contain the given port after
the address, separated by a colon.

Notice that because the port is resolved dynamically together with the
address, all logging output references the port as given, which may be a
service name (e.g. http).