service update: avoid deadlock by returning resolver context earlier
service_lookup_thread() calls service_update(), which in turn might need
to wait for additional name service lookups also requiring a context.

This could lead to an additional delay until contexts became available
or even deadlocks.

We avoid this situation by transferring ownership of the resolver
context to service_update() to return it before waiting for more
resolutions to complete.

The issue was identified, correctly analyzed and the fix was suggested
by @jake-dog, thank you.

Fixes #78