Set version 2.7.0, and update RPM packaging for Varnish 7.1.2.
The packaging script now runs yum update to bring the OS in the
rpmbuild container up to date. This is a good idea in any case,
and is now necessary for updates of the ca-certificates package,
because otherwise TLS connections to package repos fail. But it
does mean that a packaging build takes considerably longer.

Thanks to @janosmiko for contributing updates (minor tweaks in
this commit).

Ref #93