FIX backend reuse for via
This fixes a regression introduced with
75546cf550e5bffe75fccd241e77206f126fbff2:

With via, we can not compare the backends' endpoints, because they point
to the via backend and the actual information is in the proxy header.

I do not want to re-introduce a container struct for backends, so for
now we just dup the suckaddr in the reference.