Add the authority parameter to .backend() with a director() default
Previously the authority was always set to the value of the
VRT_backend host header. That is still the default, but with the new
parameter a different value may be set, or the parameter may be set to
the empty string, to specify that no authority TLV should be sent.

This commit is based on a previous implementation by Geoff Simmons
@slimhazard, who also wrote most of the test case and the commit
message. Thank you!

Closes #76