Keep our backends until the direcor is destroyed
8f582a792015399326c442cd6c16e8105fc53269 missed one aspect:

In order to support the case where our director is referenced by
another director for longer than our own timeout, we can not release
out own backends when we return our own reference. We need to keep our
backends around until the last reference to a domain is lost,
otherwise it would stop working.

We should be safe to do this because we do not add arbitraty
references to other directors, but rather we create all backends which
we use solely through this director and are thus free to release them
whenever we like.