Another workaround for VC#4037
To avoid VRT_AddDirector() tripping on the "Dynamic Backends can only be
added to warm VCLs" assertion, we need to ensure that we hold a VCL
reference until all resolver threads are stopped, such that the VCL
temperature remains VCL_COOLING and VRT_AddDirector() returns NULL
instead of panicking.

Yet because of the wrong COLD event order in VC#4037, we run into a
catch-22:

- we can not return the reference until all directors received a cold
  event, which only happens after the VMOD cold event,
- but we can neither keep the reference until they are all done, because
  that trips another assertion in mgt that a cold event needs to
  succeed.

Because we do not want to keep this legacy in the main branch, we add
the workarounds specifically for 7.4.

Ref https://github.com/varnishcache/varnish-cache/pull/4037

Closes #108